Fix PST/UTC problem with admin analytics functional tests for AB#15796 (#5296)
