+++ Apps/Admin/Tests/Functional/cypress/integration/e2e/communications.cy.js
@@ -1,28 +1,4 @@
-function getTodayPlusDaysDate(days) {
-    let newDay = new Date(Date.now());
-    cy.log(`New day: ${newDay}`);
-    newDay.setDate(newDay.getDate() + days);
-    cy.log(`New day plus ${days}: ${newDay}`);
-
-    // Convert the date and time to a localized string
-    const localizedDateString = newDay.toLocaleString("en-US", {
-        timeZone: "America/Vancouver",
-    });
-
-    // Parse the localized string to a Date object
-    const parsedDate = new Date(localizedDateString);
-    cy.log(`Parsed date: ${parsedDate}`);
-
-    // Get the components: year, month, and day
-    const year = parsedDate.getFullYear();
-    const month = String(parsedDate.getMonth() + 1).padStart(2, "0");
-    const day = String(parsedDate.getDate()).padStart(2, "0");
-
-    // Format the date to "yyyy-mm-dd" format
-    const dateString = `${year}-${month}-${day}`;
-    cy.log(`Today plus ${days} => ${dateString}`);
-    return dateString;
-}
+import { getTodayPlusDaysDate } from "../../utilities/sharedUtilities";
 
 describe("Communications", () => {
     beforeEach(() => {
